Rename getPartnerSupportItems file
